Complete Code Overhaul
Nothing worked in the original build, so I decided to fix EVERYTHING
*The program starts by running the SubmitJob method in UserComputeImpl
*The SubmitJob takes in an DataSource to retrieve the file with the data, a DataDestination to write the computed data to and an optional delimiter .
*Submit Job takes in the data from the source through StorageComputeAPI's readData which formats each entry as a individual number in an integer list and stores it in a local inputData variable.
*inputData is computed through ComputationAPI's compute method which takes every entry, translates it into a letter, then applies the Caesar cipher to decode the actual letter. After all entries have been decoded, returns all letters into a String with no delimiters.
*Finally, submitJob will take the returned string from the compute method and add it to the writeData method along with the dataDestination and a custom delimiter if there was one.
*writeData will format the string of letters into another string that is separated by the delimiter and written into a file specified at the location of dataDestination.

*The tests now check for valid inputs, exceptions, error handling

+++ src/compute/ComputeRequest.java
@@ -1,5 +1,5 @@
 package compute;
 
 public interface ComputeRequest {
-	long getInputData();
+	int getInputData();
 }
